+++ resources/sql/ampache.sql
@@ -17,10 +17,10 @@
 -- version 5.2.1
 -- https://www.phpmyadmin.net/
 --
--- Host: 192.168.1.9
--- Generation Time: Jun 06, 2024 at 05:50 AM
--- Server version: 10.11.6-MariaDB-0+deb12u1
--- PHP Version: 8.2.18
+-- Host: 192.168.1.20
+-- Generation Time: Jul 10, 2024 at 09:15 AM
+-- Server version: 10.11.6-MariaDB-0+deb12u1-log
+-- PHP Version: 8.2.21
 
 SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
 START TRANSACTION;
@@ -33,7 +33,7 @@ SET time_zone = "+00:00";
 /*!40101 SET NAMES utf8mb4 */;
 
 --
--- Database: `ampache6`
+-- Database: `ampache7`
 --
 
 -- --------------------------------------------------------
@@ -796,6 +796,7 @@ CREATE TABLE IF NOT EXISTS `playlist` (
   `last_duration` int(11) UNSIGNED NOT NULL DEFAULT 0,
   `username` varchar(128) CHARACTER SET utf8mb3 COLLATE utf8mb3_unicode_ci DEFAULT NULL,
   `last_count` int(11) DEFAULT NULL,
+  `collaborate` varchar(255) DEFAULT NULL,
   PRIMARY KEY (`id`),
   KEY `name` (`name`),
   KEY `type` (`type`)
@@ -899,7 +900,7 @@ CREATE TABLE IF NOT EXISTS `preference` (
   UNIQUE KEY `preference_UN` (`name`),
   KEY `category` (`category`),
   KEY `name` (`name`)
-) ENGINE=InnoDB AUTO_INCREMENT=189 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
+) ENGINE=InnoDB AUTO_INCREMENT=196 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
 --
 -- Dumping data for table `preference`
@@ -1032,8 +1033,14 @@ INSERT INTO `preference` (`id`, `name`, `value`, `description`, `level`, `type`,
 (184, 'bookmark_latest', '0', 'Only keep the latest media bookmark', 25, 'boolean', 'options', NULL),
 (185, 'jp_volume', '0.8', 'Default webplayer volume', 25, 'special', 'streaming', 'player'),
 (186, 'perpetual_api_session', '0', 'API sessions do not expire', 100, 'boolean', 'system', 'backend'),
-(187, 'home_recently_played_all', '0', 'Show all media types in Recently Played', 25, 'bool', 'interface', 'home'),
-(188, 'show_wrapped', '0', 'Enable access to your personal \"Spotify Wrapped\" from your user page', 25, 'bool', 'interface', 'privacy');
+(188, 'show_wrapped', '0', 'Enable access to your personal \"Spotify Wrapped\" from your user page', 25, 'bool', 'interface', 'privacy'),
+(189, 'sidebar_hide_switcher', '0', 'Hide sidebar switcher arrows', 25, 'boolean', 'interface', 'home'),
+(190, 'sidebar_hide_browse', '0', 'Hide the Browse menu in the sidebar', 25, 'boolean', 'interface', 'home'),
+(191, 'sidebar_hide_dashboard', '0', 'Hide the Dashboard menu in the sidebar', 25, 'boolean', 'interface', 'home'),
+(192, 'sidebar_hide_video', '0', 'Hide the Video menu in the sidebar', 25, 'boolean', 'interface', 'home'),
+(193, 'sidebar_hide_search', '0', 'Hide the Search menu in the sidebar', 25, 'boolean', 'interface', 'home'),
+(194, 'sidebar_hide_playlist', '0', 'Hide the Playlist menu in the sidebar', 25, 'boolean', 'interface', 'home'),
+(195, 'sidebar_hide_information', '0', 'Hide the Information menu in the sidebar', 25, 'boolean', 'interface', 'home');
 
 -- --------------------------------------------------------
 
@@ -1483,7 +1490,7 @@ CREATE TABLE IF NOT EXISTS `update_info` (
 --
 
 INSERT INTO `update_info` (`key`, `value`) VALUES
-('db_version', '600073'),
+('db_version', '700008'),
 ('Plugin_Last.FM', '000005');
 
 -- --------------------------------------------------------
@@ -1603,6 +1610,19 @@ CREATE TABLE IF NOT EXISTS `user_playlist` (
   KEY `user` (`user`),
   KEY `object_type` (`object_type`),
   KEY `object_id` (`object_id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
+
+-- --------------------------------------------------------
+
+--
+-- Table structure for table `user_playlist_map`
+--
+
+DROP TABLE IF EXISTS `user_playlist_map`;
+CREATE TABLE IF NOT EXISTS `user_playlist_map` (
+  `playlist_id` int(11) UNSIGNED NOT NULL,
+  `user_id` int(11) UNSIGNED NOT NULL,
+  UNIQUE KEY `playlist_user` (`playlist_id`,`user_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
 -- --------------------------------------------------------
@@ -1739,7 +1759,6 @@ INSERT INTO `user_preference` (`user`, `preference`, `value`) VALUES
 (-1, 171, ''),
 (-1, 172, '0'),
 (-1, 173, '1'),
-(-1, 174, '0'),
 (-1, 175, '0'),
 (-1, 176, '0'),
 (-1, 177, '1'),
@@ -1752,8 +1771,14 @@ INSERT INTO `user_preference` (`user`, `preference`, `value`) VALUES
 (-1, 184, '0'),
 (-1, 185, '0.8'),
 (-1, 186, '0'),
-(-1, 187, '0'),
-(-1, 188, '0');
+(-1, 188, '0'),
+(-1, 189, '0'),
+(-1, 190, '0'),
+(-1, 191, '0'),
+(-1, 192, '0'),
+(-1, 193, '0'),
+(-1, 194, '0'),
+(-1, 195, '0');
 
 -- --------------------------------------------------------