+++ src/Config/functions.php
@@ -939,17 +939,25 @@ function show_tvshow_season_select($name, $season_id, $allow_add = false, $video
  * @param int $catalog_id
  * @param string $style
  * @param bool $allow_none
- * @param string $filter_type
+ * @param string $gather_types
+ * @param string $catalog_type
  */
-function show_catalog_select($name, $catalog_id, $style = '', $allow_none = false, $filter_type = ''): void
+function show_catalog_select($name, $catalog_id, $style = '', $allow_none = false, $gather_types = '', $catalog_type = ''): void
 {
     echo "<select name=\"$name\" style=\"$style\">\n";
 
     $params = [];
     $sql    = "SELECT `id`, `name` FROM `catalog` ";
-    if (!empty($filter_type)) {
-        $sql .= "WHERE `gather_types` = ? ";
-        $params[] = $filter_type;
+    $where  = "WHERE";
+    if (!empty($gather_types)) {
+        $sql .= $where . " `gather_types` = ? ";
+        $params[] = $gather_types;
+        $where    = "AND";
+    }
+
+    if (!empty($catalog_type)) {
+        $sql .= $where . " `catalog_type` = ? ";
+        $params[] = $catalog_type;
     }
     $sql .= "ORDER BY `name`;";
     $db_results = Dba::read($sql, $params);
@@ -961,9 +969,13 @@ function show_catalog_select($name, $catalog_id, $style = '', $allow_none = fals
     if ($allow_none) {
         echo "\t<option value=\"-1\">" . T_('None') . "</option>\n";
     }
-    if (empty($results) && !empty($filter_type)) {
+
+    if (
+        empty($results) &&
+        !empty($gather_types)
+    ) {
         /* HINT: Requested object string/id/type ("album", "myusername", "some song title", 1298376) */
-        echo "\t<option value=\"\" selected=\"selected\">" . sprintf(T_('Not Found: %s'), $filter_type) . "</option>\n";
+        echo "\t<option value=\"\" selected=\"selected\">" . sprintf(T_('Not Found: %s'), $gather_types) . "</option>\n";
     }
 
     foreach ($results as $row) {