+++ src/Module/Util/Ui.php
@@ -931,7 +931,7 @@ class Ui implements UiInterface
                 echo "</select>\n";
                 break;
             case 'upload_catalog':
-                show_catalog_select('upload_catalog', $value, '', true, 'music');
+                show_catalog_select('upload_catalog', $value, '', true, 'music', 'local');
                 break;
             case 'play_type':
                 $is_stream     = '';