+++ docs/CHANGELOG.md
@@ -15,13 +15,16 @@ Information and changes for this major release are recorded in the wiki [here](h
   * Add `npm_binary_path`
   * Remove OpenID config
   * Add `database_engine` to allow you to change from InnoDB if you want to
-* Database 700006
+* Database 700009
   * Add user preferences to show/hide menus in the sidebar and the switcher arrows
   * Add Indexes to `object_count`, `object_count_idx_count_type_date_id` and `object_count_idx_count_type_id`
   * Convert the remaining MyISAM tables to InnoDB
   * Drop and recreate `tmp_browse` to allow InnoDB conversion
   * Add a `last_count` to playlist table to speed up access requests
   * Delete user preference `home_recently_played_all`
+  * Add `collaborate` to the playlist table to allow other users to add songs to the list
+  * Create `user_playlist_map` table to allow browse access to playlists with collaborators
+  * Convert system preference `upload_catalog` into a user preference
 
 ### Changed
 
@@ -56,6 +59,10 @@ Information and changes for this major release are recorded in the wiki [here](h
 * Add maxlength to the user data fields on the registration and edit pages
 * Use `autocomplete="new-password"` to stop trying to autocomplete on edit pages
 * Default theme fallback to 'Dark' color scheme
+* Add users as a playlist collaborator (These users can add and remove songs from a playlist)
+* Database 600075
+  * Add a `collaborate` to playlist table to allow other users to add songs to the list
+  * Create `user_playlist_map` to allow browse access to playlists with collaborate access
 * CLI
   * admin:updateDatabase: allow downgrading database from Ampache7 databases
 
@@ -65,6 +72,8 @@ Information and changes for this major release are recorded in the wiki [here](h
 * Simplify and speed up playlist access checks by sending the user object
 * User IP history defaults to grouped values and 100 results
 * Don't notify that items are added in the WebUI when nothing was added
+* Users will see additional playlists if they are added as a collaborator
+* Don't notify that items are added to a playlist in the WebUI when nothing happened
 
 ### Removed
 
@@ -79,6 +88,7 @@ Information and changes for this major release are recorded in the wiki [here](h
 * URL encode name parameters for links
 * User IP history page didn't show what you expected and was very confusing
 * Edit search and playlist pages not reloading the user list
+* Missing table check for database update 500006
 
 ## Ampache 6.5.0